WiP Commit
just in case of power failure.
back in the dark ages, before many of you were born. maybe before your parents were born. i was programming something on an old computer. it used a tape drive. and i was near finished with testing it. eleven hours of programming, testing, and debugging. and there was a power brownout. the computer reset. i had not save because it was a hassle to save often with tape drives.
that also explains why i repair everything in every MMO i play every time i see a repair.

what does this have to do with the changes?
nothing. i just was feel i should say something.