switch from search & replace to parsing binary files [...]
Previously the primary method of finding strings to translate was to search
for occurences of the strings within the binary files, which could easily
lead to false positives and incomplete matches.

Now the files are parsed by known structures and strings are extracted
based upon headers that indicate their exact length; and known separators.

There is no guarantee that the strings found are actually used like that
ingame (they can be code segments or identifiers) but they're guaranteed to
be complete, and have a high likelihood to not be binary garbage.